Modifications because of entity changes.
git-svn-id: https://develop.eng.uts.edu.au/projects/personal/sahara/svn/SchedServer/trunk@591 7ee224fa-57ca-4419-87f9-1f4314938d77

+++ Permissions/src/au/edu/uts/eng/remotelabs/schedserver/permissions/impl/UserClassAdmin.java
@@ -79,7 +79,7 @@ public class UserClassAdmin
     {
         UserClass userClass = new UserClass();
         userClass.setName(name);
-        userClass.setPriority(priority);
+        userClass.setPriority((short) priority);
         userClass.setActive(active);
         userClass.setQueuable(queuable);
         userClass.setKickable(kickable);
@@ -122,7 +122,7 @@ public class UserClassAdmin
             return false;
         }
         
-        userClass.setPriority(priority);
+        userClass.setPriority((short) priority);
         userClass.setActive(active);
         userClass.setQueuable(queuable);
         userClass.setKickable(kickable);
@@ -168,7 +168,7 @@ public class UserClassAdmin
             userClass.setName(name);
         }
         
-        userClass.setPriority(priority);
+        userClass.setPriority((short) priority);
         userClass.setActive(active);
         userClass.setQueuable(queuable);
         userClass.setKickable(kickable);
